Global dummy receipt ficture
